Do not build packages while `get`ting them
This is a hotfix for a problem caused when `go get` automatically
building the dependency it pulls down. The latest version (e.g., master
branch) will be compiled and used instead of the source version locked
to a previous checkout.

A future commit might build the dependency using the locked version, but
that will require more changes. For now, we are just fixing the bug.